git mv to match class name
